requirements.txt
python-dateutil
requests
spotipy
python-dotenv